fixing log request middleware
